requirements.txt
uvicorn==0.23.2
fastapi==0.104.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]
